add some feature

+++ app/Http/Controllers/RegisterUserController.php
@@ -29,61 +29,105 @@ class RegisterUserController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     *
+     * @author samnag
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function register(Request $request)
-    {
-        ///set all field are required
-        $validator = Validator::make($request->all(), [
-            'email'    => 'required|email|unique:users',
+{
+    ///set all field are required
+    $validator = Validator::make($request->all(), [
+        'email'    => 'required|email|unique:users',
+    ]);
+
+    //if validation = false show message error
+    if($validator->fails()){
+        return response()->json(array('status' => 'fail','errors'=>$validator->errors()));
+    }else{
+
+        //  $photo = $request->file('image');
+        // $destinationPath = 'images/users/'; // path to save to, has to exist and be writeable
+        // $filename = $photo->getClientOriginalName(); // original name that it was uploaded with
+        // $photo->move($destinationPath,$filename); // moving the file to specified dir with the original name
+
+        $user = new Users();
+        $user->username = $request->input('username');
+        $user->email = $request->input('email');
+        $user->password = sha1($request->input('password')); //encrypt password
+        //$user->image = $filename;
+        $user->status = 1;
+        $user->save();
+        //response message
+        return response()->json(array('status'=> 'success','users' => $user));
+    }
+}
+
+
+    /**
+     * This method is used for buyer
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function login(Request $request){
+        $validator = Validator::make($request->all(), [//check validation required
+            'email' => 'required|email',
+            'password' => 'required',
         ]);
-
-        //if validation = false show message error
-        if($validator->fails()){
-            return response()->json(array('status' => 'fail','errors'=>$validator->errors()));
+        if ($validator->fails()) {
+            return response()->json(array('status' => 'fail','errors'=>$validator->errors()));//return message error
         }else{
-            $photo = $request->file('image');
-            $destinationPath = 'images/users/'; // path to save to, has to exist and be writeable
-            $filename = $photo->getClientOriginalName(); // original name that it was uploaded with
-            $photo->move($destinationPath,$filename); // moving the file to specified dir with the original name
-
-            $user = new Users();
-            $user->username = $request->input('username');
-            $user->email = $request->input('email');
-            $user->password = sha1($request->input('password')); //encrypt password
-            $user->image = $filename;
-            $user->status = 1;
-            $user->save();
-
-            //response message
-            return response()->json(array('status'=> 'success','users' => $user));
+            $email = $request->email;
+            $password = $request->password;
+            $login = DB::select('
+                select * from users where users.email = "'.$email.'" and users.password = "'.sha1($password).'"
+            ');
+            if(count($login) > 0){//check is true or not
+                return response()->json(array(
+                    'status'=>"success",
+                    'sms'=> 'Login successfully!!',
+                    'data'=>$login
+                ));
+            }else{
+                return response()->json(array(
+                    'status'=> "fail",
+                    'sms'=> 'Login not success, Please try again!!'
+                ));
+            }
         }
     }
 
+    /**
+     * This email is used to display user info in their profile
+     * @author Chhin
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function userProfile($id)
     {
-        $users = DB::table('users')
-            ->select('users.id','users.username','users.image','users.email','users.password','users.address')
-            ->where('users.id',$id)->get();
-        if($users){
-            return response()->json(array('status' => 'success', 'posterProfile' => $users));
+        $userData = Users::find($id);
+        if($userData){
+            return response()->json(array('status' => 'success', 'posterProfile' => $userData));
         }else{
             return response()->json(array(
                 'status' => 'fail','message' =>'No record', ),200);
         }
     }
+
+    /**
+     * This method is used to view user's favorite
+     * @author Chhin
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function viewUserFavorite($id)
     {
-       $user = DB::table('users')
-           ->join("favorites", "users.id", "=", "favorites.users_id")
+       $user = \DB::table('favorites')
+           ->select('favorites.users_id as userId','favorites.posts_id as postId','posts.pos_image','posts.pos_title')
            ->join("posts", "posts.id", "=", "favorites.posts_id")
-           ->select('posts.id','posts.posters_id','posts.pos_image','posts.pos_title')
-           ->where('users.id',$id)
+           ->where('favorites.users_id','=',$id)
            ->get();
         if($user){
-            return response()->json(array('status' => 'success', 'users' => $user,));
+            return response()->json(array('status' => 'success', 'users' => $user));
         }else{
             return response()->json(array(
                 'status' => 'fail','message' =>'No record', ),200);
@@ -101,9 +145,11 @@ class RegisterUserController extends Controller
 
     public function changeCover(Request $request,$id){
         $userID = Users::find($id);
-
+        $oldCover = $userID->covers;
         if($request->file('covers')) {
-            File::delete('images/users/'.$userID->covers);
+            if($oldCover != "dj.png"){
+                File::delete('images/users/'.$userID->covers);
+            }
 
             $image = $request->file('covers');
             $fileName = $image->getClientOriginalName();
@@ -130,9 +176,11 @@ class RegisterUserController extends Controller
 
         $userID = Users::find($id);
         //$userID = new Users();
-
+        $oldProfile = $userID->image;
         if($request->file('image')) {
-            File::delete('images/users/'.$userID->image);
+            if($oldProfile != "dj.png"){
+                File::delete('images/users/'.$oldProfile);
+            }
 
             $image = $request->file('image');
             $fileName = $image->getClientOriginalName();
